Revert "[Front-End] Add Navigation"
