+++ front-end/package.json
@@ -19,9 +19,7 @@
     "react-dom": "^16.8.6",
     "react-native": "https://github.com/expo/react-native/archive/sdk-34.0.0.tar.gz",
     "react-native-elements": "^1.2.0",
-    "react-native-web": "^0.11.4",
-    "react-navigation": "^4.0.5",
-    "react-navigation-stack": "^1.7.3"
+    "react-native-web": "^0.11.4"
   },
   "devDependencies": {
     "@types/jest": "^24.0.18",
