+++ front-end/src/App.tsx
@@ -1,9 +1,19 @@
-import React, { Component } from 'react';
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
 
-import AppNavigator from './configs/routes/Navigator';
-
-export default class App extends Component {
-  render() {
-    return <AppNavigator />;
-  }
+export default function App() {
+  return (
+    <View style={styles.container}>
+      <Text>Open up App.tsx to start working on your app!</Text>
+    </View>
+  );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
