--- front-end/src/configs/routes/Navigator.ts
@@ -1,18 +1,0 @@
-import { createSwitchNavigator, createAppContainer } from 'react-navigation';
-
-import AppStack from './AppStack';
-import AuthStack from './AuthStack';
-import LoginScene from '../../scenes/LoginScene';
-
-const AppNavigator = createSwitchNavigator(
-  {
-    App: AppStack,
-    Auth: AuthStack,
-    Login: LoginScene,
-  },
-  {
-    initialRouteName: 'Login',
-  },
-);
-
-export default createAppContainer(AppNavigator);
